requirements.txt
python-telegram-bot
pandas
openpyxl
pyyaml
